requirements.txt
python-telegram-bot==13.5
urllib3==1.26.15
environs==11.2.0
google-cloud-dialogflow
google-cloud-api-keys
